requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
base58==2.1.1
bitarray==2.6.0
cached-property==1.5.2
cattrs==22.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cytoolz==0.12.0
decorator==5.1.1
eth-abi==2.2.0
eth-account==0.5.8
eth-bloom==1.0.4
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-tester==0.6.0b6
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.0.0rc9
executing==0.9.1
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
importlib-resources==5.9.0
ipfshttpclient==0.8.0a2
ipython==8.4.0
jedi==0.18.1
jsonschema==4.9.0
lru-dict==1.1.8
matplotlib-inline==0.1.3
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.30
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-ecc==5.2.0
py-evm==0.5.0a3
py-geth==3.8.0
pycryptodome==3.15.0
pyethash==0.1.27
Pygments==2.12.0
pyrsistent==0.18.1
pysha3==1.0.2
python-dotenv==0.20.0
requests==2.28.1
requests-cache==0.9.6
rlp==2.0.1
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.3.0
toolz==0.12.0
traitlets==5.3.0
trie==2.0.0a5
typing-extensions==3.10.0.2
url-normalize==1.4.3
urllib3==1.26.11
varint==1.0.2
wcwidth==0.2.5
web3==5.30.0
websockets==9.1
yarl==1.7.2
zipp==3.8.1
